requirements.txt
bambu-connect==0.3.1
bambulabs_api==2.5.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
paho-mqtt==2.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
